refactor: move prisma setup into create templates
