+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "create-prisma",
   "version": "0.1.3",
-  "description": "Create and initialize Prisma 7 projects with first-party templates and great DX.",
+  "description": "Create Prisma 7 projects with first-party templates and great DX.",
   "repository": {
     "type": "git",
     "url": "https://github.com/prisma/create-prisma"
