+++ src/index.ts
@@ -2,12 +2,9 @@ import { os } from "@orpc/server";
 import { createCli } from "trpc-cli";
 
 import { runCreateCommand } from "./commands/create";
-import { runInitCommand } from "./commands/init";
 import {
   CreateCommandInputSchema,
-  InitCommandInputSchema,
   type CreateCommandInput,
-  type InitCommandInput,
 } from "./types";
 
 const CLI_VERSION = process.env.CREATE_PRISMA_CLI_VERSION ?? "0.0.0";
@@ -23,15 +20,6 @@ export const router = os.router({
     .handler(async ({ input }) => {
       await runCreateCommand(input ?? {});
     }),
-  init: os
-    .meta({
-      description: "Initialize Prisma in your current project",
-      negateBooleans: true,
-    })
-    .input(InitCommandInputSchema.optional())
-    .handler(async ({ input }) => {
-      await runInitCommand(input ?? {});
-    }),
 });
 
 export function createCreatePrismaCli() {
@@ -42,21 +30,16 @@ export function createCreatePrismaCli() {
   });
 }
 
-export async function init(input: InitCommandInput = {}): Promise<void> {
-  await runInitCommand(input);
-}
-
 export async function create(input: CreateCommandInput = {}): Promise<void> {
   await runCreateCommand(input);
 }
 
-export type { CreateCommandInput, InitCommandInput };
+export type { CreateCommandInput };
 export {
   CreateCommandInputSchema,
   CreateTemplateSchema,
   DatabaseProviderSchema,
   DatabaseUrlSchema,
-  InitCommandInputSchema,
   PackageManagerSchema,
   SchemaPresetSchema,
 } from "./types";
