+++ templates/create/hono/README.md.hbs
@@ -60,7 +60,26 @@ Generated by `create-prisma` with the Hono template.
 {{/if}}
 {{/if}}
 {{/if}}
+4. Seed the database:
+
+{{#if (eq packageManager "bun")}}
+`bun run db:seed`
+{{else}}
+{{#if (eq packageManager "pnpm")}}
+`pnpm db:seed`
+{{else}}
+{{#if (eq packageManager "npm")}}
+`npm run db:seed`
+{{else}}
+`npm run db:seed` or `pnpm db:seed` or `bun run db:seed`
+{{/if}}
+{{/if}}
+{{/if}}
+
+5. Use the Prisma client from `src/lib/prisma.ts`.
+
+Generated Prisma files are written to `generated/prisma`.
 {{#if (eq schemaPreset "basic")}}
 
-The template includes a basic `User` model and a sample `GET /users` endpoint.
+The template includes a basic `User` model, a sample `GET /users` endpoint, and seed data in `prisma/seed.ts`.
 {{/if}}
