+++ templates/create/hono/prisma/schema.prisma.hbs
@@ -1,6 +1,6 @@
 generator client {
   provider = "prisma-client"
-  output   = "./generated"
+  output   = "../generated/prisma"
 }
 
 datasource db {
