+++ templates/create/next/README.md.hbs
@@ -31,15 +31,18 @@ Generated by `create-prisma` with the Next.js template.
 Prisma setup is scaffolded automatically in:
 
 - `prisma/schema.prisma`
-- `prisma/index.ts`
+- `prisma/seed.ts`
+- `src/lib/prisma.ts`
 - `prisma.config.ts`
+- `generated/prisma`
 
 Database helper scripts are added to `package.json`:
 
 - `db:generate`
 - `db:push`
 - `db:migrate`
+- `db:seed`
 {{#if (eq schemaPreset "basic")}}
 
-The starter page reads from a basic `User` model so you can verify queries quickly.
+The starter page reads from a basic `User` model so you can verify queries quickly, and `prisma/seed.ts` inserts starter users.
 {{/if}}
