+++ templates/create/next/app/page.tsx.hbs
@@ -1,5 +1,5 @@
 {{#if (eq schemaPreset "basic")}}
-import { prisma } from "../prisma";
+import { prisma } from "../src/lib/prisma";
 {{/if}}
 
 export default async function Home() {
