+++ templates/create/next/package.json.hbs
@@ -20,6 +20,7 @@
     "@types/react-dom": "^19",
     "eslint": "^9",
     "eslint-config-next": "16.1.6",
+    "tsx": "^4.7.1",
     "typescript": "^5"
   }
 }
